requirements.txt
# Flask framework
Flask

# Helper lib for running scripts locally without server
flask-script

# SCSS Compiler
Flask-Scss

# Javascript Compiler
Flask-Assets

# YAML load/compiler
PyYAML

# URL library for handling external resources
requests
